WIP reflectToMemory
