+++ src/prompts/tableAanalysis.ts
@@ -1,0 +1,103 @@
+export default `
+Tables and Fields
+tbFuHandelse
+This table seems to record events or transactions related to facilities. Key fields include:
+
+recHandelse (int): Primary key, unique identifier for each event.
+intKundnr (int): Customer number, linking the event to a customer.
+strAnlnr (varchar): Facility number, linking the event to a facility.
+intTjanstnr (int): Service number, indicating the type of service.
+datDatumFrom and datDatumTom (datetime): Start and end dates of the event.
+strTaxekod (varchar): Tariff code, linking to the tariff table.
+lngFakturanr (bigint): Invoice number, linking the event to an invoice.
+Various other fields related to the event details, such as strRFID, strFordon, strAvvikelsekod, and strHandelseText.
+tbFuRenhAvvikelse
+This table records deviations or issues related to events.
+
+intRecnum (int): Primary key, unique identifier for each deviation.
+strAvvikelsekod (varchar): Deviation code, primary key.
+strAvvikelsetext (varchar): Text describing the deviation.
+Various boolean fields indicating different statuses and actions related to the deviation.
+tbFuTaxa
+This table contains tariff or pricing information.
+
+intRecnum (int): Primary key, unique identifier for each tariff.
+strTaxekod (varchar): Tariff code, primary key.
+strTaxebenamning (varchar): Tariff name.
+Various fields related to pricing, grouping, and applicability of the tariff.
+tbFuAnlaggning
+This table contains information about facilities.
+
+intRecnum (int): Primary key, unique identifier for each facility.
+strAnlnr (varchar): Facility number, primary key.
+intKundnr (int): Customer number, linking the facility to a customer.
+Various fields related to the address, categorization, and additional details of the facility.
+Relationships
+tbFuHandelse references tbFuAnlaggning through strAnlnr.
+tbFuHandelse references tbFuTaxa through strTaxekod.
+tbFuHandelse references tbFuRenhAvvikelse through strAvvikelsekod.
+Key Insights
+Facility Management: The tbFuAnlaggning table holds details about each facility, which are linked to events in tbFuHandelse.
+Event Tracking: The tbFuHandelse table is central to tracking all events, services, and transactions for each facility.
+Deviations Handling: The tbFuRenhAvvikelse table helps manage and document any issues or deviations related to events.
+Tariff Management: The tbFuTaxa table provides detailed tariff information, crucial for billing and pricing.
+
+Database Structure Analysis for tbFuTaxa
+
+The tbFuTaxa table in the provided database contains the following fields:
+
+strTaxekod:
+Description: This field likely represents the unique code or identifier for each tax type or service fee.
+Data Type: String
+Purpose: It serves as the primary key or unique identifier for each record in the table. It helps in distinguishing different types of taxes or service fees.
+strTaxebenamning:
+Description: This field describes the name or the details of the tax or service fee.
+Data Type: String
+Purpose: It provides a descriptive name for each tax type or service fee, allowing users to understand the nature of the fee or tax.
+Insights and Observations:
+
+Diversity of Tax Codes:
+
+The table contains a wide range of tax codes (e.g., 0001, 0002, 1101, 1201, 240K104), suggesting it covers various tax types and service fees related to different aspects such as water (e.g., VA, V), waste (e.g., RESTAVFALL, KÄRL), and administrative fees.
+Naming Conventions:
+
+The strTaxekod uses a combination of letters and numbers which appear to follow specific conventions. For example:
+VA likely stands for water-related services.
+V might refer to specific types of water charges.
+K and S in the middle or end could indicate different types of containers or services (e.g., KÄRL for bins or containers, SÄCK for bags).
+Service Frequency:
+
+The tax descriptions often include numbers that seem to indicate the frequency of the service (e.g., 13 ggr/år, 52 ggr/år which translate to 13 times a year, 52 times a year). This provides insights into how often these services are billed or provided.
+Specific Services:
+
+Some codes are very specific, detailing particular services or scenarios (e.g., Slamtömning for sludge emptying, Sprinkleranläggning for sprinkler systems, Matavfall for food waste collection). This suggests the database is used for managing a detailed and comprehensive range of municipal services.
+Administrative and Miscellaneous Charges:
+
+There are entries for administrative tasks and other charges (e.g., ADMAVG for administrative fees, DRÖJRÄNTA for penalty interest). This indicates the table also includes financial penalties and additional service-related fees.
+Comprehensive Coverage:
+
+The table appears to comprehensively cover various aspects of municipal services including water supply, sewage, waste management, and administrative processes.
+Use Cases:
+
+Billing and Invoicing:
+
+The table can be used to generate bills and invoices for residents and businesses based on the services they use.
+Service Management:
+
+It helps in managing different municipal services, scheduling collections, and ensuring all provided services are documented and billed correctly.
+Reporting and Analysis:
+
+The table allows for detailed reporting and analysis of tax revenues, service usage, and operational efficiency of municipal services.
+Suggestions for Improvement:
+
+Normalization:
+
+If the database grows, consider normalizing the tables to reduce redundancy. For example, separate tables for service types, frequencies, and container types might help in managing the data more efficiently.
+Additional Fields:
+
+Adding fields such as effective_date and expiration_date for each tax code could help manage changes over time more effectively.
+Data Validation:
+
+Implement data validation rules to ensure the consistency and accuracy of the codes and descriptions.
+By understanding the structure and content of the tbFuTaxa table, one can better manage and utilize the database for municipal services and financial management.
+`
