+++ src/workers/index.ts
@@ -4,3 +4,4 @@ export { default as summarizeAsk } from './summarizeAsk'
 export { default as planAnswer } from './planAnswer'
 export { default as dataFetcher } from './dataFetcher'
 export { default as answerQuestion } from './answerQuestion'
+export { default as reflectToMemory } from './reflectToMemory'
